vijay|#00|fixed redirects in usercontroller

+++ ananya-bbc-reports-users/src/test/java/org/motechproject/ananya/bbc/users/service/UserServiceTest.java
@@ -106,7 +106,7 @@ public class UserServiceTest {
 
         UserView userResponse = userService.getUser(user.getId());
 
-        assertEquals((int) user.getId(), userResponse.getId());
+        assertEquals(user.getId(), userResponse.getId());
         assertEquals(user.getName(), userResponse.getName());
         assertEquals(user.getUsername(), userResponse.getUsername());
